+++ spec/unit/hanami/view/version_spec.rb
@@ -1,5 +1,5 @@
 RSpec.describe "Hanami::View::VERSION" do
-  it 'exposes version' do
-    expect(Hanami::View::VERSION).to eq('1.0.1')
+  it "exposes version" do
+    expect(Hanami::View::VERSION).to eq("1.1.0")
   end
 end